move camera to its own module
